Frozen-Bubble with real playlist support
I am a fan of the game Frozen-Bubble but I found the playlist option to be misleading in that it could only play all files in a given directory. I modified the code and added the option of real playlists that are text files that simply list the audio files (including the path) that you want to be played, for example playlists from XMMP or Winamp should work. Incase you didn't know the majority of Frozen-Bubble's code is Perl. You can dl the version with my updates added by the original developer (Guillaume) here.